Extract PythonRunner from PythonRDD.scala file.
